+++ packages/components/babel.config.js
@@ -1,0 +1,8 @@
+/**
+ * Babel config
+ *
+ * @since [version]
+ * @version [version]
+ */
+
+module.exports = { presets: [ '@babel/preset-react' ] };