+++ packages/components/src/index.js
@@ -1,0 +1,5 @@
+import ButtonGroupControl from './button-group-control';
+
+export {
+	ButtonGroupControl,
+};